Continue refactoring Header
